Time: 961 ms (13.43%), Space: 271.4 MB (24.07%) - LeetHub

+++ 0327-count-of-range-sum/0327-count-of-range-sum.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
-    void merge(vector<int>&nums, int l, int m, int r, int& lower, int& upper, int& count){
+    void merge(vector<long long>&nums, int l, int m, int r, int& lower, int& upper, int& count){
         int sz1 = m - l + 1;
         int sz2 = r - m ;
-        vector<int>v1(sz1),v2(sz2);
+        vector<long long>v1(sz1),v2(sz2);
         
         for(int i=l;i<=r;i++){
             if(i <= m)v1[i-l] = nums[i];
@@ -39,7 +39,7 @@ public:
         }
     }
 
-    void mergeSort(vector<int>&nums, int l, int r, int& lower, int& upper, int& count){
+    void mergeSort(vector<long long>&nums, int l, int r, int& lower, int& upper, int& count){
         if(l>=r) return ;
         int m = (l+r)>>1;
         mergeSort(nums, l, m, lower, upper, count);
@@ -50,7 +50,7 @@ public:
     int countRangeSum(vector<int>& nums, int lower, int upper) {
         int n = nums.size();
         int count =0;
-        vector<int>pre(n+1, 0);
+        vector<long long>pre(n+1, 0);
         for(int i=0;i<n;i++){
             pre[i+1] = nums[i];
             if(i>0)pre[i+1] += pre[i];
